<feature> 修改了中断读取数据中的bug,现在已经可以正常读取中断数据

+++ lib/include/bmi055_driver.hpp
@@ -20,6 +20,6 @@ class BMI055
 
 
 
-    int m_acc_accd_x_mg,m_acc_accd_y_mg,m_acc_accd_z_mg;
+    float m_acc_accd_x_mg,m_acc_accd_y_mg,m_acc_accd_z_mg;
 
 };
